The relationship between loan and employee was removed, entity status was added in LoanResponse, field, customer and book status was added in Loanrequest, employee status was added in Request

+++ src/main/java/com/examplenewstack/newstack/client/controller/api/DeleteClientByIdController.java
@@ -18,7 +18,7 @@ public class DeleteClientByIdController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<?> deleteById(@RequestBody ClientRequestDTO clientRequestDTO, @PathVariable Long id) {
+    public ResponseEntity<?> deleteById(@RequestBody ClientRequestDTO clientRequestDTO, @PathVariable int id) {
 
         deleteClientByIdService.deleteById(id);
         return ResponseEntity.ok().build();
