+++ src/main/java/com/examplenewstack/newstack/client/repository/ClientRepository.java
@@ -9,7 +9,8 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClientRepository extends JpaRepository<Client, Long> {
+public interface ClientRepository extends JpaRepository<Client, Integer> {
+
     Client findByCPFAndPassword(String cpf, String password);
 
     boolean findByCPF(String cpf);
