+++ src/main/java/com/examplenewstack/newstack/client/service/ReportsClientByIdService.java
@@ -21,7 +21,7 @@ public class ReportsClientByIdService {
     }
 
     public ClientResponseDTO showClientById(
-            Long id
+            int id
     ){
         Optional<Client> client = clientRepository.findById(id);
         if (client.isEmpty()) {
